refactor: restructure models by moving Task, Team, and User classes to separate files
