+++ core/models/__init__.py
@@ -1,4 +1,6 @@
-from .entities import Task, Team, User
 from .enums import TaskPriority, TaskStatus
+from .task import Task
+from .team import Team
+from .user import User
 
 __all__ = ["Task", "Team", "User", "TaskPriority", "TaskStatus"]
